feat(visits): implement update visit status functionality and remove deprecated route
