+++ src/controller/visits.client.controller.js
@@ -119,32 +119,6 @@ export const updateVisit = async (req, res, next) => {
     }
 }
 
-export const updateVisitStatus = async (req, res, next) => {
-    const { id } = req.params
-    const { status } = req.body
-
-    try {
-
-        const visit = await Visit.findByIdAndUpdate(id, { status }, { new: true }).select("-client -staff -address -type -notes")
-
-        if (status === "cancelled") {
-            const { cancellationReason } = req.body
-            visit.cancellationReason = cancellationReason
-            await visit.save()
-        }
-
-        return res.status(200).json({
-            status: true,
-            message: "Visit status updated successfully",
-            data: visit
-        })
-    }
-
-    catch (error) {
-        next(error)
-    }
-}
-
 export const updateVisitNotes = async (req, res, next) => {
 
     const { id } = req.params
